+++ src/core/params/env-variables.ts
@@ -1,8 +1,8 @@
 import { ArgumentsParser } from "../../cli/ArgumentsParser";
+import { BuidlerArguments, BuidlerParamDefinitions } from "../../types";
 import { unsafeObjectKeys } from "../../util/unsafe";
 import { BuidlerError, ERRORS } from "../errors";
 
-import { BuidlerArguments, BuidlerParamDefinitions } from "./buidler-params";
 import ProcessEnv = NodeJS.ProcessEnv;
 
 const BUIDLER_ENV_ARGUMENT_PREFIX = "BUIDLER_";
